added Utilities.jsx, transferred some functions
